Add versioning to Gulpfile.js
This commit adds versioning to the gulpfile, eliminating
the caching problems on some systems by rewriting the
paths to these files every compilation.

This has some problem where the compilation takes a bit
longer and making the CSS file unaccessible for a while
The browser won't render the CSS if it is unavailable.
This is fixed by reloading after a second or waiting for
the Elixir popup to come.

Signed-off-by: MegaXLR <admin@megaxlr.net>